Update design.md
Hello, after changing the name of the file, I can not edit the file in intial pull request. I am not sure if you are in the same case as me. 
So to make sure all of us can open it, I created a new pull request for design.md file.
If you have any new comments, just let me know! Thank you!