Update design.md
Hello, after changing the name of designmd file.
I can not edit the file in the initial pull request. I am not sure that you are in the same case as me.
So I created a new pull request to make sure all of us can open it.
If you have any new comments, just let me know. Thank you!